+++ dart/geobase/lib/geobase.dart
@@ -42,6 +42,7 @@ export 'src/common/constants/screen_ppi.dart';
 export 'src/common/conversions/angle_unit.dart';
 export 'src/common/conversions/area_unit.dart';
 export 'src/common/conversions/distance_unit.dart';
+export 'src/common/conversions/time_unit.dart';
 export 'src/common/functions/geographic_functions.dart';
 export 'src/common/functions/position_functions.dart';
 export 'src/common/presentation/dms.dart';